# apis/api/swagger/swagger.yaml
components:
...
    schemas:
...
        Tool:
            properties:
                _id:
                    type: string
                description:
                    type: string
                    description: 'description of the Tool'
                name:
                    type: string
                    description: 'The name of the Tool'
            required:
                - _id
                - description
                - name
            type: object
        Pick_Tool.Exclude_keyofTool._id:
            properties:
                name:
                    type: string
                    description: 'The name of the Tool'
                description:
                    type: string
                    description: 'description of the Tool'
            required:
                - name
                - description
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        OmitOverride_Tool._id_:
            $ref: '#/components/schemas/Pick_Tool.Exclude_keyofTool._id'
        ToolCreate:
            $ref: '#/components/schemas/OmitOverride_Tool._id_'
        GROUPS:
            enum:
                - admin
                - user
                - owner
                - unathenticated
            type: string
        User:
            properties:
                _id:
                    type: string
                    description: 'The users mongodb id'
                phoneValidationCode:
                    type: string
                    description: 'Code texted to the user to validate their phone number'
                phone:
                    type: number
                    format: double
                    description: "the phone number of the user that can be used for phone based\r\nauthentication"
                password:
                    type: string
                    description: 'The users password, encrypted'
                email:
                    type: string
                    description: 'Email for the user'
                currentToken:
                    type: string
                    description: 'Current token for the user'
                googleProvider:
                    properties:
                        token:
                            type: string
                        id:
                            type: string
                    required:
                        - token
                        - id
                    type: object
                    description: 'data from Google for google SSO'
                sessionStartDate:
                    type: string
                    format: date-time
                    description: 'The date and time that the users session began'
                passwordResetExpires:
                    type: string
                    format: date-time
                    description: 'date at which the password hash is no longer valid for resetting'
                passwordResetHash:
                    type: string
                    description: 'password reset hash that the user can use to reset their password'
                userPermissions:
                    items:
                        $ref: '#/components/schemas/GROUPS'
                    type: array
                    description: 'Permission level of the user'
            required:
                - _id
                - userPermissions
            type: object
        Pick_User.email-or-phone-or-password_:
            properties:
                email:
                    type: string
                    description: 'Email for the user'
                phone:
                    type: number
                    format: double
                    description: "the phone number of the user that can be used for phone based\r\nauthentication"
                password:
                    type: string
                    description: 'The users password, encrypted'
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        UserCreate:
            $ref: '#/components/schemas/Pick_User.email-or-phone-or-password_'
    securitySchemes:
        jwt:
            type: apiKey
            name: authorization
            in: header
info:
    title: '@helium/ts-tsoa-api'
    version: 0.0.2
    description: 'Helium API with TSOA generator and openAPI sdk generator'
    contact: {}
openapi: 3.0.0
paths:
...
    /tool/find/all:
        post:
            operationId: AdminFindtool
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Tool'
                                type: array
            tags:
                - tool
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /tool/find:
        post:
            operationId: Findtools
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Tool'
                                type: array
            tags:
                - tool
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/tool/{toolId}':
        get:
            operationId: Gettool
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tool'
            tags:
                - tool
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: toolId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updatetool
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tool'
            tags:
                - tool
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: toolId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ToolCreate'
        delete:
            operationId: Deletetool
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - tool
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: toolId
                    required: true
                    schema:
                        type: string
    /tool:
        post:
            operationId: Createtool
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tool'
            tags:
                - tool
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ToolCreate'
    /user/find/all:
        post:
            operationId: AdminFinduser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/User'
                                type: array
            tags:
                - user
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /user/find:
        post:
            operationId: Findusers
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/User'
                                type: array
            tags:
                - user
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/user/{userId}':
        get:
            operationId: Getuser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
            tags:
                - user
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: userId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updateuser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
            tags:
                - user
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: userId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserCreate'
        delete:
            operationId: Deleteuser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - user
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: userId
                    required: true
                    schema:
                        type: string
    /user:
        post:
            operationId: Createuser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
            tags:
                - user
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserCreate'
servers:
    -
        url: 'http://localhost:3002/v1'